7.1 lambda text modifier

+++ kodilla-testing/src/main/java/calculator/Calculator.java
@@ -2,10 +2,10 @@ package calculator;
 
 public class Calculator {
 
-    public int calculatorAdd(int a, int b) {
-        return a+b;
+    public Double calculatorAdd(double a, double b) {
+        return a + b;
     }
-    public int calculatorSubtract(int a, int b) {
+    public Double calculatorSubtract(double a, double b) {
         return a-b;
     }
 }
